+++ src/builtin/modules/random.cpp
@@ -13,7 +13,7 @@
  *
  * Author: Zhenjie Wei
  * Created: Jul. 29, 2025
- * Updated: Dec. 11, 2025
+ * Updated: Dec. 24, 2025
  * Supported by: National Key Research and Development Program of China
  */
 
@@ -28,7 +28,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
             {
                 {
                     ":rand/seed",
-                    StaticFuncTypeResolver::create({}, {{Type::Int(), false}}, Type::Void()),
+                    StaticFuncTypeResolver::create({}, {{Type::Int64(), false}}, Type::Void()),
                 },
             }),
         OperatorGroup::create(
@@ -36,7 +36,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
             {
                 {
                     ":rand/rand",
-                    StaticFuncTypeResolver::create({}, {}, Type::Double()),
+                    StaticFuncTypeResolver::create({}, {}, Type::Float64()),
                 },
             }),
         OperatorGroup::create(
@@ -44,7 +44,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
             {
                 {
                     ":rand/randn",
-                    StaticFuncTypeResolver::create({}, {}, Type::Double()),
+                    StaticFuncTypeResolver::create({}, {}, Type::Float64()),
                 },
             }),
         OperatorGroup::create(
@@ -54,8 +54,8 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                     ":rand/randint",
                     StaticFuncTypeResolver::create(
                         {},
-                        {{Type::Int(), false}, {Type::Int(), false}},
-                        Type::Int()),
+                        {{Type::Int64(), false}, {Type::Int64(), false}},
+                        Type::Int64()),
                 },
             }),
         OperatorGroup::create(
