+++ src/builtin/modules/tensor.cpp
@@ -13,7 +13,7 @@
  *
  * Author: Zhenjie Wei
  * Created: Oct. 21, 2025
- * Updated: Dec. 11, 2025
+ * Updated: Dec. 24, 2025
  * Supported by: National Key Research and Development Program of China
  */
 
@@ -34,7 +34,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                     ":tensor/ones",
                     StaticFuncTypeResolver::create(
                         {},
-                        {{ArrayType::create(Type::Int()), false}},
+                        {{ArrayType::create(Type::Int64()), false}},
                         TensorType::Default()),
                 },
             }),
@@ -45,7 +45,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                     ":tensor/zeros",
                     StaticFuncTypeResolver::create(
                         {},
-                        {{ArrayType::create(Type::Int()), false}},
+                        {{ArrayType::create(Type::Int64()), false}},
                         TensorType::Default()),
                 },
             }),
@@ -56,7 +56,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                     ":tensor/eye",
                     StaticFuncTypeResolver::create(
                         {},
-                        {{Type::Int(), false}},
+                        {{Type::Int64(), false}},
                         TensorType::Default()),
                 },
             }),
@@ -78,7 +78,9 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                     ":tensor/linspace",
                     StaticFuncTypeResolver::create(
                         {},
-                        {{Type::Float(), false}, {Type::Float(), false}, {Type::Float(), false}},
+                        {{Type::Float32(), false},
+                         {Type::Float32(), false},
+                         {Type::Float32(), false}},
                         TensorType::Default()),
                 },
             }),
@@ -92,10 +94,11 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                         "(start: float, stop: float, step?: float) => Tensor",
                         [](const type_vec_t &with, const type_vec_t &norm, const ModifierSet &)
                             -> optional<type_ptr_t> {
-                            if (norm[0]->equals(Type::Float()) && norm[1]->equals(Type::Float())) {
+                            if (norm[0]->equals(Type::Float32()) &&
+                                norm[1]->equals(Type::Float32())) {
                                 if (norm.size() == 2) {
                                     return TensorType::Default();
-                                } else if (norm.size() == 3 && norm[2]->equals(Type::Float())) {
+                                } else if (norm.size() == 3 && norm[2]->equals(Type::Float32())) {
                                     return TensorType::Default();
                                 }
                             }
@@ -115,12 +118,12 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                             -> optional<type_ptr_t> {
                             if (norm[0]->code() == TypeCode::Array) {
                                 auto arrType = tt::as_shared<ArrayType>(norm[0]);
-                                if (arrType->elemType()->equals(Type::Int())) {
+                                if (arrType->elemType()->equals(Type::Int64())) {
                                     if (norm.size() == 1) {
                                         return TensorType::Default();
                                     } else if (norm.size() == 3) {
-                                        if (norm[1]->equals(Type::Float()) &&
-                                            norm[2]->equals(Type::Float())) {
+                                        if (norm[1]->equals(Type::Float32()) &&
+                                            norm[2]->equals(Type::Float32())) {
                                             return TensorType::Default();
                                         }
                                     }
@@ -142,12 +145,12 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                             -> optional<type_ptr_t> {
                             if (norm[0]->code() == TypeCode::Array) {
                                 auto arrType = tt::as_shared<ArrayType>(norm[0]);
-                                if (arrType->elemType()->equals(Type::Int())) {
+                                if (arrType->elemType()->equals(Type::Int64())) {
                                     if (norm.size() == 1) {
                                         return TensorType::Default();
                                     } else if (norm.size() == 3) {
-                                        if (norm[1]->equals(Type::Float()) &&
-                                            norm[2]->equals(Type::Float())) {
+                                        if (norm[1]->equals(Type::Float32()) &&
+                                            norm[2]->equals(Type::Float32())) {
                                             return TensorType::Default();
                                         }
                                     }
@@ -168,7 +171,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                         [](const type_vec_t &with, const type_vec_t &norm, const ModifierSet &)
                             -> optional<type_ptr_t> {
                             if (norm[0]->code() == TensorType::typeCode()) {
-                                auto int32Type = Type::Int();
+                                auto int32Type = Type::Int64();
                                 return std::make_shared<ArrayType>(int32Type);
                             }
                             return nullopt;
@@ -191,16 +194,16 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                                 return norm[0];
                             }
                             if (norm[0]->code() == TensorType::typeCode() &&
-                                (norm[1]->code() == TypeCode::Int ||
-                                 norm[1]->code() == TypeCode::Long ||
-                                 norm[1]->code() == TypeCode::Float ||
-                                 norm[1]->code() == TypeCode::Double)) {
+                                (norm[1]->code() == TypeCode::Int32 ||
+                                 norm[1]->code() == TypeCode::Int64 ||
+                                 norm[1]->code() == TypeCode::Float32 ||
+                                 norm[1]->code() == TypeCode::Float64)) {
                                 return norm[0];
                             }
-                            if ((norm[0]->code() == TypeCode::Int ||
-                                 norm[0]->code() == TypeCode::Long ||
-                                 norm[0]->code() == TypeCode::Float ||
-                                 norm[0]->code() == TypeCode::Double) &&
+                            if ((norm[0]->code() == TypeCode::Int32 ||
+                                 norm[0]->code() == TypeCode::Int64 ||
+                                 norm[0]->code() == TypeCode::Float32 ||
+                                 norm[0]->code() == TypeCode::Float64) &&
                                 norm[1]->code() == TensorType::typeCode()) {
                                 return norm[1];
                             }
@@ -223,16 +226,16 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                                 return norm[0];
                             }
                             if (norm[0]->code() == TensorType::typeCode() &&
-                                (norm[1]->code() == TypeCode::Int ||
-                                 norm[1]->code() == TypeCode::Long ||
-                                 norm[1]->code() == TypeCode::Float ||
-                                 norm[1]->code() == TypeCode::Double)) {
+                                (norm[1]->code() == TypeCode::Int32 ||
+                                 norm[1]->code() == TypeCode::Int64 ||
+                                 norm[1]->code() == TypeCode::Float32 ||
+                                 norm[1]->code() == TypeCode::Float64)) {
                                 return norm[0];
                             }
-                            if ((norm[0]->code() == TypeCode::Int ||
-                                 norm[0]->code() == TypeCode::Long ||
-                                 norm[0]->code() == TypeCode::Float ||
-                                 norm[0]->code() == TypeCode::Double) &&
+                            if ((norm[0]->code() == TypeCode::Int32 ||
+                                 norm[0]->code() == TypeCode::Int64 ||
+                                 norm[0]->code() == TypeCode::Float32 ||
+                                 norm[0]->code() == TypeCode::Float64) &&
                                 norm[1]->code() == TensorType::typeCode()) {
                                 return norm[1];
                             }
@@ -257,12 +260,12 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                                 return norm[0];
                             }
                             if (lhs == TensorType::typeCode() &&
-                                (rhs == TypeCode::Int || rhs == TypeCode::Long ||
-                                 rhs == TypeCode::Float || rhs == TypeCode::Double)) {
+                                (rhs == TypeCode::Int32 || rhs == TypeCode::Int64 ||
+                                 rhs == TypeCode::Float32 || rhs == TypeCode::Float64)) {
                                 return norm[0];
                             }
-                            if ((lhs == TypeCode::Int || lhs == TypeCode::Long ||
-                                 lhs == TypeCode::Float || lhs == TypeCode::Double) &&
+                            if ((lhs == TypeCode::Int32 || lhs == TypeCode::Int64 ||
+                                 lhs == TypeCode::Float32 || lhs == TypeCode::Float64) &&
                                 rhs == TensorType::typeCode()) {
                                 return norm[1];
                             }
@@ -286,16 +289,16 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                                 return norm[0];
                             }
                             if (norm[0]->code() == TensorType::typeCode() &&
-                                (norm[1]->code() == TypeCode::Int ||
-                                 norm[1]->code() == TypeCode::Long ||
-                                 norm[1]->code() == TypeCode::Float ||
-                                 norm[1]->code() == TypeCode::Double)) {
+                                (norm[1]->code() == TypeCode::Int32 ||
+                                 norm[1]->code() == TypeCode::Int64 ||
+                                 norm[1]->code() == TypeCode::Float32 ||
+                                 norm[1]->code() == TypeCode::Float64)) {
                                 return norm[0];
                             }
-                            if ((norm[0]->code() == TypeCode::Int ||
-                                 norm[0]->code() == TypeCode::Long ||
-                                 norm[0]->code() == TypeCode::Float ||
-                                 norm[0]->code() == TypeCode::Double) &&
+                            if ((norm[0]->code() == TypeCode::Int32 ||
+                                 norm[0]->code() == TypeCode::Int64 ||
+                                 norm[0]->code() == TypeCode::Float32 ||
+                                 norm[0]->code() == TypeCode::Float64) &&
                                 norm[1]->code() == TensorType::typeCode()) {
                                 return norm[1];
                             }
@@ -337,16 +340,16 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                                 return norm[0];
                             }
                             if (norm[0]->code() == TensorType::typeCode() &&
-                                (norm[1]->code() == TypeCode::Int ||
-                                 norm[1]->code() == TypeCode::Long ||
-                                 norm[1]->code() == TypeCode::Float ||
-                                 norm[1]->code() == TypeCode::Double)) {
+                                (norm[1]->code() == TypeCode::Int32 ||
+                                 norm[1]->code() == TypeCode::Int64 ||
+                                 norm[1]->code() == TypeCode::Float32 ||
+                                 norm[1]->code() == TypeCode::Float64)) {
                                 return norm[0];
                             }
-                            if ((norm[0]->code() == TypeCode::Int ||
-                                 norm[0]->code() == TypeCode::Long ||
-                                 norm[0]->code() == TypeCode::Float ||
-                                 norm[0]->code() == TypeCode::Double) &&
+                            if ((norm[0]->code() == TypeCode::Int32 ||
+                                 norm[0]->code() == TypeCode::Int64 ||
+                                 norm[0]->code() == TypeCode::Float32 ||
+                                 norm[0]->code() == TypeCode::Float64) &&
                                 norm[1]->code() == TensorType::typeCode()) {
                                 return norm[1];
                             }
@@ -366,7 +369,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                         [](const type_vec_t &with, const type_vec_t &norm, const ModifierSet &)
                             -> optional<type_ptr_t> {
                             if (norm[0]->code() == TensorType::typeCode() &&
-                                norm[1]->equals(Type::Int())) {
+                                norm[1]->equals(Type::Int64())) {
                                 auto tensor_type = std::dynamic_pointer_cast<TensorType>(norm[0]);
                                 if (tensor_type && tensor_type->shape().size() == 2 &&
                                     tensor_type->shape()[0] == tensor_type->shape()[1]) {
@@ -438,7 +441,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                         {},
                         {{TensorType::Default(), false},
                          {TensorType::Default(), false},
-                         {Type::Int(), false}},
+                         {Type::Int64(), false}},
                         TensorType::Default()),
                 },
             }),
@@ -465,7 +468,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                         [](const type_vec_t &with, const type_vec_t &norm, const ModifierSet &)
                             -> optional<type_ptr_t> {
                             if (norm[0]->code() == TensorType::typeCode()) {
-                                return Type::Float();
+                                return Type::Float32();
                             }
                             return nullopt;
                         }),
@@ -542,7 +545,7 @@ static const std::vector<oper_group_ptr_t> &getOperatorGroups() {
                     StaticFuncTypeResolver::create(
                         {},
                         {{TensorType::Default(), false}},
-                        TensorType::create(Type::Double(), {0})),
+                        TensorType::create(Type::Float64(), {0})),
                 },
             }),
         OperatorGroup::create(
